refactor: Refactor RunTaskModal children components
No jira.

This aims to split SystemsSelect component into smaller components (TaskName, TaskDescription), that would have their own separated state and business logic.
